Renamed and added relevant functions for tournament creation
